+++ src/front/js/pages/Reviews.js
@@ -1,11 +1,12 @@
 import React from 'react'
+import CardsReview from '../component/CardsReview'
 
 const Reviews = () => {
 
 
   return (
     <div>
-      <ReviewCard />
+      <CardsReview />
     </div>
   )
 }
